feat: Update the project to work with Xcode 8.0 and Swift 3.0
The app now builds and runs in Xcode 8.0 and has been tested
on a device running iOS 10.  Manual changes to the code were few:

- 	URL.fileURL(withPathComponents: pathArray) was changed to
	NSURL.fileURL(withPathComponents: pathArray)

- Calls to try! when working with the audio player and audio recorder
	in AVFoundation now use do-try-catch blocks

There is a remaining issue, namely that the update to Xcode has
changed how things look in storyboard.  This does not affect the app
when it is running (the layout is still nice) but the layout when viewed
in storyboard does not accurately reflect where the UI components
will be located when the app is in use.